pyproject.toml
[project]
name = "alisto"
version = "0.1.0"
description = ""
authors = [
    {name = "IndigonaBlue",email = "rdadeveyra@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard] (>=0.121.1,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
